tests: split "tests/test_host.py" from "tests/test_common.py"
"host.py" is (potentially) used by "common.py", and never the other way
around. As such, it would be useful to have the tests for "host.py"
independent from the tests for "common.py".

Split.
